Refactor: Extract business logic from gui.rs into separate modules
- Created gui_controller.rs to handle all business logic and state management
- Created gui_renderer.rs to handle UI rendering logic
- Simplified gui.rs to focus only on the main app structure and event handling
- Separated concerns: controller handles data/logic, renderer handles UI
- Maintained all existing functionality while improving code organization
- Fixed borrowing issues with proper state management
- Made code more maintainable and testable
