+++ src/bin/gui.rs
@@ -1,1078 +1,66 @@
 use eframe::egui;
 use sie_generate_config::{
-    backend::{
-        opcua_poller::{OpcUaNode, OpcUaPoller},
-        ConfigGenerator, ServiceType,
-    },
-    error::{TelegrafError, XmlFileValidation},
-    TelegrafConfig, WorkerCommand, WorkerHandle, WorkerResponse,
+    gui_controller::GuiController,
+    gui_renderer::GuiRenderer,
 };
 
-#[derive(Default)]
-struct XmlFileConfig {
-    namespace: String,
-    interval_ms: String,
-    ip: String,
-}
-
-#[derive(Default, Debug, PartialEq, Clone)]
-enum OpcUaBrowseState {
-    #[default]
-    Idle,
-    BrowsingNodes,
-    BrowsingNodesFailed(String),
-    BrowsingNodesComplete,
-    GettingNamespaces,
-    GettingNamespacesFailed(String),
-    GettingNamespacesComplete,
-}
-
-#[derive(Default, Debug)]
-struct FormState {
-    show_namespace_error: bool, // Track if we should show namespace errors
-    show_iot_host_error: bool,  // Track if IOT host is invalid
-    ip_errors: std::collections::HashMap<String, bool>, // Track file IP validation errors
-}
-
 struct TelegrafApp {
-    config: TelegrafConfig,
-    xml_files: Vec<String>,
-    selected_listener_files: Vec<bool>, // Checkboxes for listener selection
-    file_configs: std::collections::HashMap<String, XmlFileConfig>,
-    status_messages: Vec<String>, // Changed from status_message to store multiple messages
-    form_state: FormState,               // Validation state for the form
-    opcua_nodes: Vec<OpcUaNode>,         // Store the complete OPC UA node hierarchy
-    show_opcua_browser: bool,            // Toggle for showing the OPC UA browser
-    browse_status_message: String,       // Status message for OPC UA browsing
-    worker: Option<WorkerHandle>,        // Background worker for async operations
-    is_working: bool,                    // Whether a background operation is in progress
-    opcua_browse_state: OpcUaBrowseState, // State for OPC UA browsing operations
-    should_scroll: bool, // Whether to scroll to bottom on next render
+    controller: GuiController,
 }
 
 impl TelegrafApp {
-    /// Helper method to send a command to the worker and update UI state
-    fn send_worker_command(&mut self, command: WorkerCommand, working_message: &str) {
-        if let Some(worker) = &self.worker {
-            if let Err(e) = worker.send_command(command) {
-                self.status_messages.push(format!("Failed to start operation: {}", e));
-            } else {
-                self.is_working = true;
-                self.status_messages.push(working_message.to_string());
-            }
-        } else {
-            self.status_messages.push("Worker not initialized".to_string());
+    fn new() -> Self {
+        Self {
+            controller: GuiController::new(),
         }
-    }
-
-    fn load_xml_files(&mut self) {
-        // Use the shared function from lib.rs
-        self.xml_files = sie_generate_config::discover_xml_files(&self.config.folder);
-        self.selected_listener_files = vec![false; self.xml_files.len()];
-
-        // Initialize configs for new files
-        for file in &self.xml_files {
-            self.file_configs.entry(file.clone()).or_default();
-        }
-    }
-
-    // Helper function to set UI error flags and get user-friendly error message
-    fn handle_error(&mut self, error: &TelegrafError, context: &str) -> String {
-        let message = error.user_friendly_message(context);
-
-        // Set UI error flags based on the error message
-        if message.contains("Host Format Error") {
-            self.form_state.show_iot_host_error = true;
-        }
-
-        // Additional flags can be set here as needed
-
-        message
-    }
-
-    // Render the OPC UA node tree recursively with support for lazy loading
-    fn render_node_tree(
-        &mut self,
-        ui: &mut egui::Ui,
-        nodes: &mut [OpcUaNode],
-        indent_level: usize,
-    ) {
-        for node in nodes.iter_mut() {
-            // Calculate indentation
-            let indent = (indent_level as f32) * 20.0; // 20 pixels per indent level
-            ui.horizontal(|ui| {
-                ui.add_space(indent);
-
-                // Determine if this is a folder-like node
-                let is_folder = node.is_folder_node();
-                
-                // Show checkboxes for variables that can be selected and for folders
-                // Skip checkbox for the root node (at indent_level 0)
-                if (node.node_class == opcua::types::NodeClass::Variable || is_folder) && indent_level > 0 {
-                    if ui.checkbox(&mut node.selected, "").changed() {
-                        // For variables: If a node is deselected, also deselect all its children
-                        if !node.selected && node.node_class == opcua::types::NodeClass::Variable {
-                            node.deselect_children();
-                        }
-                        
-                        // For folders: When selected, ensure children are loaded
-                        if node.selected && is_folder && !node.children_loaded {
-                            // Clone to avoid borrow issues
-                            let node_clone = node.clone();
-                            
-                            // Create a new poller to load children
-                            if let Ok(poller) = OpcUaPoller::new(self.config.clone()) {
-                                match poller.load_node_children(&node_clone, indent_level) {
-                                    Ok(children) => {
-                                        // Update the node with loaded children
-                                        node.children = children;
-                                        node.children_loaded = true;
-                                        
-                                        // Force a redraw
-                                        ui.ctx().request_repaint();
-                                    }
-                                    Err(e) => {
-                                        // Log the error but don't display it in the UI to avoid disrupting the layout
-                                        eprintln!("Error loading folder children when selecting checkbox: {}", e);
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-
-                // Node display - show different icons based on node type
-                let node_icon = match node.node_class {
-                    opcua::types::NodeClass::Object => "[O] ",
-                    opcua::types::NodeClass::Variable => "[V] ",
-                    opcua::types::NodeClass::Method => "[M] ",
-                    opcua::types::NodeClass::ObjectType => "[T] ",
-                    opcua::types::NodeClass::VariableType => "[VT] ",
-                    opcua::types::NodeClass::ReferenceType => "[R] ",
-                    opcua::types::NodeClass::DataType => "[D] ",
-                    opcua::types::NodeClass::View => "[~] ",
-                    _ => "[?] ",
-                };
-
-                // Check if this is a folder-like node that can have children
-                if is_folder {
-                    
-                    let label = format!("{}{} ({:?})", node_icon, node.display_name, node.node_class);
-                    
-                    // Use collapsing header to show node and its children
-                    let header = ui.collapsing(label, |ui| {
-                        // Show additional node information
-                        if let Some(data_type) = &node.data_type {
-                            ui.label(format!("Data Type: {}", data_type));
-                        }
-                        if let Some(description) = &node.description {
-                            ui.label(format!("Description: {}", description));
-                        }
-
-                        // Check if children need to be loaded
-                        if !node.children_loaded && node.children.is_empty() {
-                            // Show loading indicator
-                            ui.horizontal(|ui| {
-                                ui.spinner();
-                                ui.label("Loading children...");
-                            });
-
-                            // Clone to avoid borrow issues
-                            let node_clone = node.clone();
-
-                            // Create a new poller to load children
-                            if let Ok(poller) = OpcUaPoller::new(self.config.clone()) {
-                                match poller.load_node_children(&node_clone, indent_level) {
-                                    Ok(children) => {
-                                        // Update the node with loaded children
-                                        node.children = children;
-                                        node.children_loaded = true;
-
-                                        // Force a redraw
-                                        ui.ctx().request_repaint();
-                                    }
-                                    Err(e) => {
-                                        ui.label(format!("Error loading children: {}", e));
-                                    }
-                                }
-                            }
-                        } else {
-                            // Render children recursively
-                            self.render_node_tree(ui, &mut node.children, indent_level + 1);
-                        }
-                    });
-
-                    // If the header is expanded, we don't need to show the hover text
-                    if !header.fully_open() {
-                        header.header_response.on_hover_text(format!(
-                            "NodeId: {:?}\nNamespace: {}\nBrowse Name: {}{}",
-                            node.node_id,
-                            node.node_id.namespace,
-                            node.browse_name,
-                            if let Some(data_type) = &node.data_type {
-                                format!("\nData Type: {}", data_type)
-                            } else {
-                                String::new()
-                            }
-                        ));
-                    }
-                } else {
-                    // Leaf node without children
-                    let label =
-                        format!("{}{} ({:?})", node_icon, node.display_name, node.node_class);
-                    ui.label(label).on_hover_text(format!(
-                        "NodeId: {:?}\nNamespace: {}\nBrowse Name: {}{}",
-                        node.node_id,
-                        node.node_id.namespace,
-                        node.browse_name,
-                        if let Some(data_type) = &node.data_type {
-                            format!("\nData Type: {}", data_type)
-                        } else {
-                            String::new()
-                        }
-                    ));
-                }
-            });
-        }
-    }
-
-    // Add selected nodes to the configuration
-    fn add_selected_nodes_to_config(&mut self) {
-        // Clear any existing selections first
-        self.config.selected_opcua_nodes.clear();
-
-        // Use the new utility function to convert selected nodes to configuration
-        let selected_nodes = OpcUaNode::convert_selected_nodes_to_config(&self.opcua_nodes);
-        self.config.selected_opcua_nodes = selected_nodes;
-    }
-
-    fn default() -> Self {
-        let mut path = std::env::current_exe().unwrap();
-        path.pop(); // Remove the executable name
-
-        let worker = Some(WorkerHandle::new());
-
-        let mut app = Self {
-            config: TelegrafConfig {
-                folder: path.clone(),
-                ip: env!("DEFAULT_IP").to_string(),
-                username: env!("DEFAULT_USERNAME").to_string(),
-                password: env!("DEFAULT_PASSWORD").to_string(),
-                iot_host: env!("DEFAULT_IOT_IP").to_string(),
-                iot_username: env!("DEFAULT_IOT_USERNAME").to_string(),
-                iot_password: env!("DEFAULT_IOT_PASSWORD").to_string(),
-                listener_files: Vec::new(),
-                output_format: Some("influxdb".to_string()),
-                include_test_inputs: true,
-                selected_opcua_nodes: Vec::new(),
-            },
-            xml_files: Vec::new(),
-            selected_listener_files: Vec::new(),
-            file_configs: std::collections::HashMap::new(),
-            status_messages: Vec::new(), // Initialize with empty vector for status messages
-            form_state: FormState::default(),
-            opcua_nodes: Vec::new(),
-            show_opcua_browser: false,
-            browse_status_message: String::new(),
-            worker,
-            is_working: false,
-            opcua_browse_state: OpcUaBrowseState::default(),
-            should_scroll: true, // Start with auto-scroll enabled
-        };
-
-        // Load initial data
-        app.load_xml_files();
-        app
     }
 }
 
 impl eframe::App for TelegrafApp {
     fn update(&mut self, ctx: &egui::Context, _frame: &mut eframe::Frame) {
-        // Check for worker responses
-        if let Some(worker) = &self.worker {
-            // First, check if we have any responses
-            if let Some(response) = worker.try_get_response() {
-                // Set working state for non-progress responses
-                if !matches!(response, WorkerResponse::ProgressUpdate(_)) {
-                    self.is_working = false;
-                    self.should_scroll = true; // Set flag to scroll on next render
-                }
-
-                // Process the response
-                match response {
-                    WorkerResponse::DummyResponse => {
-                        self.status_messages.push("Dummy operation completed!".to_string());
-                    }
-                    WorkerResponse::SshCommandOutput(output) => {
-                        self.status_messages.push(format!("SSH command output:\n{}", output));
-                    }
-                    WorkerResponse::SshError(err) => {
-                        self.status_messages.push(format!("SSH error: {}", err));
-                    }
-                    WorkerResponse::FileTransferComplete => {
-                        self.status_messages.push("File transfer completed successfully".to_string());
-                    }
-                    WorkerResponse::FileTransferError(err) => {
-                        self.status_messages.push(format!("File transfer error: {}", err));
-                    }
-                    WorkerResponse::ProgressUpdate(progress) => {
-                        // Update or add progress message
-                        if let Some(last_msg) = self.status_messages.last_mut() {
-                            if last_msg.starts_with("Progress:") {
-                                *last_msg = progress;
-                            } else if last_msg != &progress {  // Only add if different
-                                self.status_messages.push(progress);
-                            }
-                        } else {
-                            self.status_messages.push(progress);
-                        }
-                        self.should_scroll = true; // Ensure we scroll for progress updates
-                    }
-                    WorkerResponse::OpcUaNodes(nodes) => {
-                        self.opcua_nodes = nodes;
-                        self.opcua_browse_state = OpcUaBrowseState::BrowsingNodesComplete;
-                        self.browse_status_message = if self.opcua_nodes.is_empty() {
-                            "OPC UA structure loaded, but no nodes found.".to_string()
-                        } else {
-                            "OPC UA structure loaded successfully.".to_string()
-                        };
-                        // self.is_working is already set to false above for non-progress responses
-                    }
-                    WorkerResponse::OpcUaNamespaces(namespace_map) => {
-                        let mut found_count = 0;
-                        for (file_name, namespace_index) in namespace_map {
-                            if let Some(full_path) = self
-                                .xml_files
-                                .iter()
-                                .find(|path| path.ends_with(&file_name))
-                            {
-                                if let Some(config) = self.file_configs.get_mut(full_path) {
-                                    config.namespace = namespace_index.to_string();
-                                    found_count += 1;
-                                }
-                            }
-                        }
-                        if found_count > 0 {
-                            self.status_messages.push(
-                                format!("Found namespaces for {} XML files!", found_count)
-                            );
-                        } else {
-                            self.status_messages.push("No matching namespaces found. Check XML filenames match namespace names.".to_string());
-                        }
-                        self.opcua_browse_state = OpcUaBrowseState::GettingNamespacesComplete;
-                        // self.is_working is already set to false above
-                    }
-                    WorkerResponse::OpcUaError(err) => {
-                        // General status message for any OPC UA error
-                        self.status_messages.push(format!("OPC UA operation failed: {}", err));
-                        match self.opcua_browse_state {
-                            OpcUaBrowseState::BrowsingNodes => {
-                                self.opcua_browse_state =
-                                    OpcUaBrowseState::BrowsingNodesFailed(err.clone());
-                                self.browse_status_message =
-                                    format!("Failed to browse OPC UA structure: {}", err);
-                            }
-                            OpcUaBrowseState::GettingNamespaces => {
-                                // Status_message is already set above, specific browse_status_message not needed here
-                                self.opcua_browse_state =
-                                    OpcUaBrowseState::GettingNamespacesFailed(err);
-                            }
-                            // If error occurs in other states, just log to status_message, keep browse_state as is or reset to Idle if appropriate
-                            _ => {
-                                // Potentially reset to Idle or a generic error state if the current state is not specific to an ongoing opcua op
-                                // For now, we just let status_message show the error.
-                                // If an error occurs during BrowsingNodesComplete, for example, it's likely a new, unrelated error.
-                            }
-                        }
-                        // self.is_working is already set to false above
-                    }
-                    WorkerResponse::Error(err) => {
-                        // Handle generic worker errors
-                        self.status_messages.push(format!("Worker error: {}", err));
-                        // Reset any ongoing operations
-                        self.is_working = false;
-                    }
-                }
-                // Always request a repaint when we have a response
-                ctx.request_repaint();
-            } else if self.is_working {
-                // If we're working but don't have a response yet, request a repaint
-                // to keep the spinner animating
-                ctx.request_repaint();
-            }
+        // Process worker responses and request repaint if needed
+        if self.controller.process_worker_responses() {
+            ctx.request_repaint();
         }
 
-        // Show loading indicator next to Command Output when working
+        // Main UI rendering
         egui::CentralPanel::default().show(ctx, |ui| {
             ui.heading("Telegraf Configuration Generator");
+            
             // Configuration Section
-            ui.collapsing("Configuration", |ui| {
-                // Folder selection
-                ui.horizontal(|ui| {
-                    ui.label("XML Folder:");
-                    if ui.button("Browse").clicked() {
-                        if let Some(path) = rfd::FileDialog::new().pick_folder() {
-                            self.config.folder = path.clone();
-                            // Update XML files list
-                            self.xml_files = std::fs::read_dir(&self.config.folder)
-                                .unwrap()
-                                .filter_map(|entry| {
-                                    let path = entry.ok()?.path();
-                                    if path.is_file()
-                                        && path.extension().is_some_and(|ext| ext == "xml")
-                                    {
-                                        Some(path.to_str()?.to_string())
-                                    } else {
-                                        None
-                                    }
-                                })
-                                .collect();
-                            self.selected_listener_files = vec![false; self.xml_files.len()];
-                        }
-                    }
-                    ui.label(self.config.folder.to_string_lossy().to_string());
-                });
-
-                // Main Configuration using Grid
-                egui::Grid::new("config_grid")
-                    .num_columns(2)
-                    .spacing([40.0, 4.0])
-                    .show(ui, |ui| {
-                        // IP Configuration
-                        ui.label("OPC IP:");
-                        ui.text_edit_singleline(&mut self.config.ip);
-                        ui.end_row();
-
-                        // IOT Host with validation
-                        ui.label("IOT Host:");
-                        let text_edit = egui::TextEdit::singleline(&mut self.config.iot_host);
-                        if self.form_state.show_iot_host_error {
-                            egui::Frame::NONE
-                                .stroke(egui::Stroke::new(
-                                    1.0,
-                                    egui::Color32::from_rgb(255, 0, 0),
-                                ))
-                                .show(ui, |ui| ui.add(text_edit));
-                        } else {
-                            ui.add(text_edit);
-                        }
-                        ui.end_row();
-                    });
-
-                // Test Inputs Toggle
-                ui.horizontal(|ui| {
-                    ui.label("Include System Inputs");
-                    if ui
-                        .checkbox(
-                            &mut self.config.include_test_inputs,
-                            "CPU, Disk, Memory, of the IOT device",
-                        )
-                        .changed()
-                    {
-                        // Checkbox state is automatically saved to config
-                    }
-                });
-
-                // Credentials
-                ui.collapsing("Credentials", |ui| {
-                    egui::Grid::new("credentials_grid")
-                        .num_columns(2)
-                        .spacing([40.0, 4.0])
-                        .striped(true)
-                        .show(ui, |ui| {
-                            // OPC Username
-                            ui.label("OPC Username:");
-                            ui.text_edit_singleline(&mut self.config.username);
-                            ui.end_row();
-                            // OPC Password
-                            ui.label("OPC Password:");
-                            ui.add(egui::TextEdit::singleline(&mut self.config.password).password(true));
-                            ui.end_row();
-                            // IOT Username
-                            ui.label("IOT Username:");
-                            ui.text_edit_singleline(&mut self.config.iot_username);
-                            ui.end_row();
-                            // IOT Password
-                            ui.label("IOT Password:");
-                            ui.add(egui::TextEdit::singleline(&mut self.config.iot_password).password(true));
-                            ui.end_row();
-                        });
-
-                    // Calculate is_prometheus value once
-                    let mut is_prometheus = self
-                        .config
-                        .output_format
-                        .clone()
-                        .unwrap_or_else(|| "influxdb".to_string())
-                        == "prometheus";
-
-                    // Output Format Toggle
-                    ui.separator();
-                    ui.horizontal(|ui| {
-                        ui.label("Output Format:");
-
-                        let toggle_text = if is_prometheus {
-                            "Prometheus"
-                        } else {
-                            "InfluxDB"
-                        };
-                        if ui.button(toggle_text).clicked() {
-                            is_prometheus = !is_prometheus;
-                            self.config.output_format = Some(if is_prometheus {
-                                "prometheus".to_string()
-                            } else {
-                                "influxdb".to_string()
-                            });
-                        }
-
-                        ui.label(if is_prometheus {
-                            "(exposes metrics via HTTP)"
-                        } else {
-                            "(sends to InfluxDB)"
-                        });
-                    });
-
-                    // InfluxDB token is now handled via environment variable
-                });
-            });
-
-            // XML Files Section
-            ui.heading("XML Files Configuration");
-            if self.xml_files.is_empty() {
-                ui.label("No XML files found in the selected folder");
-            } else {
-                ui.label("Configure XML files (check for listeners/subscribers):");
-                for (i, file) in self.xml_files.iter().enumerate() {
-                    ui.group(|ui| {
-                        // Get or create config for this file
-                        let file_config = self.file_configs.entry(file.clone()).or_default();
-
-                        // File name and listener checkbox
-                        ui.horizontal(|ui| {
-                            ui.checkbox(&mut self.selected_listener_files[i], "Listener");
-                            ui.strong(file);
-                        });
-
-                        // Use grid layout for all fields in the XML file configuration
-                        egui::Grid::new(&format!("xml_file_grid_{}", i))
-                            .num_columns(2)
-                            .spacing([40.0, 4.0])
-                            .show(ui, |ui| {
-                                // Namespace input
-                                ui.label("Namespace:");
-                                let text_edit = egui::TextEdit::singleline(&mut file_config.namespace);
-                                if self.form_state.show_namespace_error {
-                                    egui::Frame::NONE
-                                        .stroke(egui::Stroke::new(
-                                            1.0,
-                                            egui::Color32::from_rgb(255, 0, 0),
-                                        ))
-                                        .show(ui, |ui| ui.add(text_edit));
-                                } else {
-                                    ui.add(text_edit);
-                                }
-                                ui.end_row();
-
-                                // IP Address input
-                                ui.label("OPC IP:");
-
-                                // Check if we have a validation error for this file
-                                let has_error = self.form_state.ip_errors.get(file).unwrap_or(&false);
-
-                                // Show the field with appropriate styling
-                                let response = if *has_error {
-                                    // If there's an error, show red border
-                                    let response = egui::Frame::NONE
-                                        .stroke(egui::Stroke::new(
-                                            1.0,
-                                            egui::Color32::from_rgb(255, 0, 0),
-                                        ))
-                                        .show(ui, |ui| {
-                                            ui.add(egui::TextEdit::singleline(&mut file_config.ip)
-                                                .hint_text(&self.config.ip))
-                                        })
-                                        .inner;
-                                    response.on_hover_text("Invalid IP format. Must be four numbers 0-255 separated by dots (e.g., 192.168.1.1)")
-                                } else {
-                                    // No error, show normal text field
-                                    let response = ui.add(egui::TextEdit::singleline(&mut file_config.ip)
-                                        .hint_text(&self.config.ip));
-                                    response.on_hover_text("Override the default OPC IP address for this file")
-                                };
-
-                                // Validate IP after user types
-                                if response.changed() {
-                                    // Only validate non-empty custom IPs
-                                    if !file_config.ip.is_empty() {
-                                        // Use the backend validation logic
-                                        let validation_result = self.config.validate_ip_for_file(&file_config.ip);
-
-                                        // Update error state
-                                        self.form_state.ip_errors.insert(file.clone(), validation_result.is_err());
-                                    } else {
-                                        // Empty IP means no error (will use default)
-                                        self.form_state.ip_errors.insert(file.clone(), false);
-                                    }
-                                }
-                                ui.end_row();
-
-                                // Interval input
-                                let is_listener = self.selected_listener_files[i];
-                                let label = if is_listener {
-                                    "Sampling Interval (ms):"
-                                } else {
-                                    "Interval (ms):"
-                                };
-                                ui.label(label);
-
-                                let default_interval = if is_listener { "500" } else { "1000" };
-                                let response = ui.add(
-                                    egui::TextEdit::singleline(&mut file_config.interval_ms)
-                                        .hint_text(default_interval),
-                                );
-
-                                response.on_hover_text(if is_listener {
-                                    "Default: 500ms for listeners"
-                                } else {
-                                    "Default: 1000ms for regular files"
-                                });
-                                ui.end_row();
-                        });
-                    });
-                    ui.add_space(4.0);
-                }
-            }
-
-            // Bucket Configuration - Only show when using InfluxDB
-            // Reuse the already calculated is_prometheus value
-            // Since it might have changed with the toggle, get the current value
-            let _is_prometheus = self
-                .config
-                .output_format
-                .clone()
-                .unwrap_or_else(|| "influxdb".to_string())
-                == "prometheus";
-
+            GuiRenderer::render_configuration(ui, &mut self.controller);
+            
+            // XML Files Configuration
+            GuiRenderer::render_xml_files_configuration(ui, &mut self.controller);
+            
             // Main Action Buttons
-            ui.horizontal(|ui| {
-                ui.horizontal(|ui| {
-                    // OPC UA Browser button
-                    if ui.button("Browse OPC UA Structure").clicked() {
-                        self.show_opcua_browser = true;
-                        // Allow triggering browse if not currently browsing nodes and not already complete from a previous run
-                        if self.opcua_browse_state != OpcUaBrowseState::BrowsingNodes && self.opcua_browse_state != OpcUaBrowseState::BrowsingNodesComplete {
-                            self.opcua_nodes.clear(); // Clear previous nodes for a fresh browse
-                            self.opcua_browse_state = OpcUaBrowseState::BrowsingNodes;
-                            self.browse_status_message = "Requesting OPC UA structure...".to_string();
-                            let command = WorkerCommand::BrowseOpcUaNodes {
-                                config: self.config.clone(),
-                            };
-                            self.send_worker_command(command, "Browsing OPC UA structure...");
-                        } else if self.opcua_browse_state == OpcUaBrowseState::BrowsingNodesComplete {
-                            // If already complete, just show the browser without re-fetching, unless user explicitly refreshes later
-                            self.browse_status_message = "OPC UA structure previously loaded.".to_string();
-                        }
-                    }
-                
-                    if ui.button("Get OPC UA Namespaces").clicked() {
-                        // Allow triggering if not currently getting namespaces
-                        if self.opcua_browse_state != OpcUaBrowseState::GettingNamespaces {
-                            self.opcua_browse_state = OpcUaBrowseState::GettingNamespaces;
-                            self.status_messages.push("Requesting OPC UA namespaces...".to_string());
-                            let command = WorkerCommand::GetOpcUaNamespaces {
-                                config: self.config.clone(),
-                                xml_files: self.xml_files.clone(),
-                            };
-                            self.send_worker_command(command, "Getting OPC UA namespaces...");
-                        }
-                    }
-                    
-                });
-                
-                if ui.button("Generate Config").clicked() {
-                    // Reset validation state
-                    self.form_state.show_namespace_error = false;
-                    self.form_state.show_iot_host_error = false;
-
-                    // Convert our file_configs to XmlFileValidation for backend validation
-                    let validation_configs: std::collections::HashMap<String, XmlFileValidation> =
-                        self.file_configs.iter().map(|(file, config)| {
-                            (file.clone(), XmlFileValidation {
-                                namespace: config.namespace.clone(),
-                                interval_ms: config.interval_ms.clone(),
-                                ip: config.ip.clone(),
-                            })
-                        }).collect();
-
-                    // Perform comprehensive backend validation
-                    match self.config.validate_config(&validation_configs) {
-                        Ok(_) => {
-                            // All validations passed
-                        },
-                        Err(errors) => {
-                            // Handle errors and update UI state
-                            let error_messages: Vec<String> = errors.iter()
-                                .map(|e| e.to_string())
-                                .collect();
-
-                            // Set appropriate error flags
-                            for error in &errors {
-                                match error {
-                                    TelegrafError::ValidationError(msg) if msg.contains("namespace") => {
-                                        self.form_state.show_namespace_error = true;
-                                    },
-                                    TelegrafError::HostFormatError(_) => {
-                                        self.form_state.show_iot_host_error = true;
-                                    },
-                                    _ => {}
-                                }
-                            }
-
-                            self.status_messages.push(format!("Validation errors: {}", error_messages.join("; ")));
-                            return;
-                        }
-                    }
-
-                    self.config.listener_files = self
-                        .xml_files
-                        .iter()
-                        .zip(self.selected_listener_files.iter())
-                        .filter(|(_, &selected)| selected)
-                        .map(|(file, _)| file.clone())
-                        .collect();
-
-                    match ConfigGenerator::new(self.config.clone()) {
-                        Ok(mut generator) => {
-                            // Set configurations for each file
-                            for file in &self.xml_files {
-                                if let Some(file_config) = self.file_configs.get(file) {
-                                    let is_listener = self.config.listener_files.contains(file);
-                                    let default_interval = if is_listener { 500 } else { 1000 };
-
-                                    let interval_ms =
-                                        file_config.interval_ms.parse().unwrap_or(default_interval);
-
-                                    // Convert empty IP string to None, otherwise Some(ip)
-                                    let ip_option = if file_config.ip.is_empty() {
-                                        None
-                                    } else {
-                                        Some(file_config.ip.clone())
-                                    };
-
-                                    generator.set_file_config(
-                                        file.clone(),
-                                        file_config.namespace.clone(),
-                                        interval_ms,
-                                        ip_option,
-                                    );
-                                }
-                            }
-                            match generator
-                                .generate_config(&self.xml_files, &self.config.listener_files)
-                            {
-                                Ok(output_path) => {
-                                    self.status_messages.push(
-                                        format!("Successfully generated config in {:?}", output_path)
-                                    );
-                                }
-                                Err(e) => {
-                                    let error_message = self.handle_error(&e, "generating config");
-                                    self.status_messages.push(error_message);
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            let error_message = self.handle_error(&e, "generating config");
-                            self.status_messages.push(error_message);
-                        }
-                    }
-                }
-
-                if ui.button("Send Config").clicked() {
-                    let config = self.config.clone();
-                    self.send_worker_command(
-                        WorkerCommand::SendTelegrafConfig { config },
-                        "Sending configuration..."
-                    );
-                }
-            });
-
+            GuiRenderer::render_main_action_buttons(ui, &mut self.controller);
+            
             // Other Commands Section
-            ui.collapsing("Other Commands", |ui| {
-                ui.horizontal(|ui| {
-                    if ui.button("Backup InfluxDB").clicked() {
-                        let config = self.config.clone();
-                        self.send_worker_command(
-                            WorkerCommand::BackupInfluxDB { config },
-                            "Backing up InfluxDB..."
-                        );
-                    }
-
-                    if ui.button("Backup Grafana").clicked() {
-                        let config = self.config.clone();
-                        self.send_worker_command(
-                            WorkerCommand::BackupGrafana { config },
-                            "Backing up Grafana..."
-                        );
-                    }
-                });
-
-                ui.horizontal(|ui| {
-                    if ui.button("Get Telegraf Status").clicked() {
-                        let config = self.config.clone();
-                        self.send_worker_command(
-                            WorkerCommand::GetTelegrafStatus { config },
-                            "Retrieving Telegraf status..."
-                        );
-                    }
-
-                    if ui.button("Get Telegraf Logs").clicked() {
-                        let config = self.config.clone();
-                        self.send_worker_command(
-                            WorkerCommand::GetTelegrafLogs { config, lines: 30 },
-                            "Retrieving Telegraf logs..."
-                        );
-                    }
-                });
-
-                // Add a new row for service status check
-                ui.horizontal(|ui| {
-                    // Calculate is_prometheus value
-                    let is_prometheus = self
-                        .config
-                        .output_format
-                        .clone()
-                        .unwrap_or_else(|| "influxdb".to_string())
-                        == "prometheus";
-
-                    // Service name based on output format
-                    let service_name = if is_prometheus { "Prometheus" } else { "InfluxDB" };
-
-                    if ui.button(format!("Check {} Status", service_name)).clicked() {
-                        let service_url = self.config.iot_host.clone();
-                        let config = self.config.clone();
-                        
-                        if is_prometheus {
-                            self.send_worker_command(
-                                WorkerCommand::CheckServiceStatus {
-                                    config,
-                                    service_url: service_url.clone(),
-                                    service_type: ServiceType::Prometheus,
-                                    timeout_secs: 5,
-                                },
-                                &format!("Checking {} status at {}...", service_name, service_url)
-                            );
-                        } else {
-                            self.send_worker_command(
-                                WorkerCommand::CheckInfluxDbStatus { config },
-                                &format!("Checking InfluxDB status at {}...", service_url)
-                            );
-                        }
-                    }
-                });
-            });
-
+            GuiRenderer::render_other_commands(ui, &mut self.controller);
+            
             // Status Messages
-            if !self.status_messages.is_empty() {
-                // Add a header to make it more visible
-                ui.separator();
-                ui.horizontal(|ui| {
-                    ui.heading("Command Output:");
-                    if self.is_working {
-                        ui.add(egui::Spinner::new().size(16.0));
-                        ui.label("Working...");
-                    }
-                    // Add a clear button
-                    if ui.button("Clear").clicked() {
-                        self.status_messages.clear();
-                        self.should_scroll = true;
-                    }
-                });
-                
-                // Create a frame with a border to make the output more visible
-                let frame = egui::Frame::dark_canvas(ui.style())
-                    .stroke(egui::Stroke::new(1.0, egui::Color32::LIGHT_BLUE));
-                    
-                frame.show(ui, |ui| {
-                    // Use scrollable area with fixed height for multiline text
-                    egui::ScrollArea::vertical()
-                        .max_height(400.0)
-                        .show(ui, |ui| {
-                            ui.style_mut().override_text_style = Some(egui::TextStyle::Monospace);
-                            
-                            // Display all messages with separators
-                            for (i, message) in self.status_messages.iter().enumerate() {
-                                if i > 0 {
-                                    ui.separator();
-                                }
-                                // Split by lines and display each line separately
-                                for line in message.lines() {
-                                    ui.label(line);
-                                }
-                            }
-                            
-                            // Only auto-scroll if we have new content and the user hasn't scrolled up
-                            if self.should_scroll {
-                                ui.allocate_rect(ui.available_rect_before_wrap(), egui::Sense::hover())
-                                    .on_hover_cursor(egui::CursorIcon::Default);
-                                
-                                // This will scroll the area to the cursor (which we just placed at the bottom)
-                                ui.scroll_to_cursor(Some(egui::Align::BOTTOM));
-                                
-                                // Reset the scroll flag after scrolling
-                                self.should_scroll = false;
-                            }
-                        });
-                    
-                    // Add status message count for reference
-                    ui.separator();
-                    ui.horizontal(|ui| {
-                        ui.label(format!("Showing {} message(s)", self.status_messages.len()));
-                    });
-                });
-            }
-
-            // OPC UA Browser Panel
-            if self.show_opcua_browser {
-                egui::Window::new("OPC UA Browser")
-                    .default_size([400.0, 600.0])
-                    .show(ctx, |ui| {
-                        match &self.opcua_browse_state {
-                            OpcUaBrowseState::BrowsingNodes => {
-                                ui.horizontal(|ui| {
-                                    ui.spinner();
-                                    ui.label(&self.browse_status_message);
-                                });
-                            }
-                            OpcUaBrowseState::BrowsingNodesComplete => {
-                                if !self.opcua_nodes.is_empty() {
-                                    let mut temp_nodes = std::mem::take(&mut self.opcua_nodes);
-                                    egui::ScrollArea::vertical().show(ui, |ui| {
-                                        self.render_node_tree(ui, &mut temp_nodes, 0);
-                                    });
-                                    self.opcua_nodes = temp_nodes;
-                                } else {
-                                    ui.label(&self.browse_status_message); // Will show "...no nodes found."
-                                }
-                            }
-                            OpcUaBrowseState::BrowsingNodesFailed(err) => {
-                                ui.colored_label(egui::Color32::RED, format!("Failed to browse OPC UA structure: {}", err));
-                            }
-                            OpcUaBrowseState::Idle => {
-                                if !self.opcua_nodes.is_empty() { // Show cached nodes if available
-                                    let mut temp_nodes = std::mem::take(&mut self.opcua_nodes);
-                                    egui::ScrollArea::vertical().show(ui, |ui| {
-                                        self.render_node_tree(ui, &mut temp_nodes, 0);
-                                    });
-                                    self.opcua_nodes = temp_nodes;
-                                } else {
-                                    ui.label("Click 'Browse OPC UA Structure' or 'Refresh Structure' to load nodes.");
-                                }
-                            }
-                            // For GettingNamespaces states, this window doesn't show specific feedback, general status is outside.
-                            _ => { 
-                                ui.label(if self.browse_status_message.is_empty() { "OPC UA Browser" } else { &self.browse_status_message });
-                            }
-                        }
-
-                        ui.separator();
-                        ui.horizontal(|ui| {
-                            if ui.button("Add Selected to Config").clicked() {
-                                self.add_selected_nodes_to_config();
-                                self.status_messages.push("Selected OPC UA nodes added to configuration.".to_string());
-                                self.should_scroll = true;
-                                // self.show_opcua_browser = false; // Keep browser open after adding
-                            }
-                            if ui.button("Refresh Structure").clicked() {
-                                 if self.opcua_browse_state != OpcUaBrowseState::BrowsingNodes {
-                                    self.opcua_nodes.clear();
-                                    self.opcua_browse_state = OpcUaBrowseState::BrowsingNodes;
-                                    self.browse_status_message = "Requesting OPC UA structure refresh...".to_string();
-                                    let command = WorkerCommand::BrowseOpcUaNodes {
-                                        config: self.config.clone(),
-                                    };
-                                    self.send_worker_command(command, "Refreshing OPC UA structure...");
-                                }
-                            }
-                            if ui.button("Close").clicked() {
-                                self.show_opcua_browser = false;
-                            }
-                        });
-                    });
-                } 
-            // Display currently selected nodes
-                    if !self.config.selected_opcua_nodes.is_empty() {
-                        ui.heading("Selected Nodes");
-                        ui.push_id("selected_nodes_area", |ui| {
-                            egui::ScrollArea::vertical().max_height(200.0).show(ui, |ui| {
-                                egui::Grid::new("selected_opcua_nodes_grid")
-                                .num_columns(5)
-                                .striped(true)
-                                .show(ui, |ui| {
-                                    ui.label("Node Name");
-                                    ui.label("Namespace");
-                                    ui.label("Measurement Name");
-                                    ui.label("Interval (ms)");
-                                    ui.label("Actions");
-                                    ui.end_row();
-                                    
-                                    let mut nodes_to_remove = Vec::new();
-                                    
-                                    for (i, node) in self.config.selected_opcua_nodes.iter_mut().enumerate() {
-                                        ui.label(&node.display_name);
-                                        ui.label(&node.namespace.to_string());
-                                        
-                                        // Allow editing the measurement name
-                                        let mut measurement_name = node.measurement_name.clone();
-                                        if ui.text_edit_singleline(&mut measurement_name).changed() {
-                                            node.measurement_name = measurement_name;
-                                        }
-                                        
-                                        // Allow editing the interval
-                                        let mut interval_str = node.interval_ms.to_string();
-                                        if ui.text_edit_singleline(&mut interval_str).changed() {
-                                            if let Ok(interval) = interval_str.parse::<u32>() {
-                                                node.interval_ms = interval;
-                                            }
-                                        }
-                                        
-                                        // Remove button
-                                        if ui.button("Remove").clicked() {
-                                            nodes_to_remove.push(i);
-                                        }
-                                        
-                                        ui.end_row();
-                                    }
-                                    
-                                    // Remove nodes that were marked for removal
-                                    for &index in nodes_to_remove.iter().rev() {
-                                        if index < self.config.selected_opcua_nodes.len() {
-                                            self.config.selected_opcua_nodes.remove(index);
-                                        }
-                                    }
-                                });
-                            });
-                        });
-                    }
-                
+            GuiRenderer::render_status_messages(ui, &mut self.controller);
+            
+            // Selected Nodes
+            GuiRenderer::render_selected_nodes(ui, &mut self.controller);
         });
+
+        // OPC UA Browser Window
+        GuiRenderer::render_opcua_browser(ctx, &mut self.controller);
     }
 }
 
 fn main() -> eframe::Result<()> {
-    // let native_options = eframe::NativeOptions {
-    //     viewport: Some(egui::vec2(800.0, 1000.0)),
-    //     ..Default::default()
-    // };
     let options = eframe::NativeOptions {
         viewport: egui::ViewportBuilder::default()
-            .with_inner_size([800.0, 1000.0]) // wide enough for the drag-drop overlay text
+            .with_inner_size([800.0, 1000.0])
             .with_drag_and_drop(true),
         ..Default::default()
     };
     eframe::run_native(
         "Telegraf Config Generator",
         options,
-        Box::new(|_cc| Ok(Box::new(TelegrafApp::default()))),
+        Box::new(|_cc| Ok(Box::new(TelegrafApp::new()))),
     )
 }
