+++ src/lib.rs
@@ -341,6 +341,8 @@ pub fn discover_xml_files(folder: &std::path::PathBuf) -> Vec<String> {
 pub mod backend;
 pub mod error;
 pub mod worker;
+pub mod gui_controller;
+pub mod gui_renderer;
 
 pub use worker::{WorkerCommand, WorkerHandle, WorkerResponse};
